Rename package in alt-common and add Vosk Test video translate to text
