+++ alt-common/src/main/java/com/alterante/utils/extractor/BinaryExtractorUtil.java
@@ -1,4 +1,4 @@
-package com.alterante.utils;
+package com.alterante.utils.extractor;
 
 import java.io.*;
 import java.nio.file.Files;
