+++ alt-common/src/main/java/com/alterante/utils/extractor/README_BinaryExtractor.md
